Update Bulgarian wordlist
Sourced from: https://github.com/dslul/openboard/pull/108

A total of approximately 750 words and their forms were added. The first line was amended with the new version and a correction for the name of the dictionary, which did not list the language name with a capitalized first letter as it should have.

The newly added words are mostly names of local dishes (cuisine), old or dialect words, Turkish loanwords that are commonplace, curses and offensive words and words pertaining to gender studies.
As is noted in the original source (https://github.com/dslul/openboard/issues/105), the new lines received the addition of ',flags=originalFreq=' for future proofing, but the old lines did not.

Other things to note if somebody wishes to improve this in the future:
1. The general frequency is skewed and needs to be heavily improved.
2. Offensive and possibly offensive flags are missing, as are any other flags that may exist.
3. Common personal and family names, famous historical figures' names, as well as monikers and/or celebrities' pseudonyms could be added to further enrich the dictionary.
4. The prefixes 'по' and 'най' (hyphenated in front of certain words) can't be handled by the dictionary as is.

Change-Id: Ibe4a1578a6d88ffbd8acc8150aa0acd3f3237010
Signed-off-by: Ray <vincentrayananda@gmail.com>